[src/si_op.c] Fixed bug in mpfr_add_si, mpfr_sub_si and mpfr_si_sub when the integer is LONG_MIN, visible when using an UB sanitizer.
[From SVN r9999 (trunk)]

+++ src/si_op.c
@@ -30,7 +30,7 @@ mpfr_add_si (mpfr_ptr y, mpfr_srcptr x, long int u, mpfr_rnd_t rnd_mode)
   if (u >= 0)
     return mpfr_add_ui (y, x, u, rnd_mode);
   else
-    return mpfr_sub_ui (y, x, -u, rnd_mode);
+    return mpfr_sub_ui (y, x, - (unsigned long) u, rnd_mode);
 }
 
 int
@@ -39,7 +39,7 @@ mpfr_sub_si (mpfr_ptr y, mpfr_srcptr x, long int u, mpfr_rnd_t rnd_mode)
   if (u >= 0)
     return mpfr_sub_ui (y, x, u, rnd_mode);
   else
-    return mpfr_add_ui (y, x, -u, rnd_mode);
+    return mpfr_add_ui (y, x, - (unsigned long) u, rnd_mode);
 }
 
 int
@@ -49,7 +49,8 @@ mpfr_si_sub (mpfr_ptr y, long int u, mpfr_srcptr x, mpfr_rnd_t rnd_mode)
     return mpfr_ui_sub (y, u, x, rnd_mode);
   else
     {
-      int res = - mpfr_add_ui (y, x, -u, MPFR_INVERT_RND (rnd_mode));
+      int res = - mpfr_add_ui (y, x, - (unsigned long) u,
+                               MPFR_INVERT_RND (rnd_mode));
       MPFR_CHANGE_SIGN (y);
       return res;
     }
